Set mappings of elastisearch to work with dini kim lrmi profile
In JSON-LD an '@context' is used to define semantics of JSON fields. These
definitions can consist of an array with mixed data types. This is invalid
when tried to be indexed into elasticsearch. As a mapping of the index can be
defined to ignore the '@context' field for indexing such an JSON-LD document
can then be loaded into elasticsearch.

The mapping must be set when initializing the index. This is done in the
start script provided by skohub-pubsub.

See https://github.com/hbz/skohub-pubsub/issues/40.